centralized all types in types.ts file
