+++ src/types/types.ts
@@ -1,0 +1,14 @@
+export type LoginInputs = {
+    email: string;
+    password: string;
+    message: string;
+    user: {}
+    success: boolean
+}
+
+export type RegisterInputs = {
+    message: string;
+    name: string;
+    email: string;
+    password: string;
+}
